% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armijo.R
\name{coarseLine}
\alias{coarseLine}
\title{Coarse Line Search is generally faster and dirtier.}
\usage{
coarseLine(fun, x, dx, beta = 3, maximise = FALSE, ...)
}
\description{
Coarse Line Search is generally faster and dirtier.
}
