% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armijo.R
\name{coordDescent}
\alias{coordDescent}
\title{Coordinate Descent}
\usage{
coordDescent(fun, x, maximise = FALSE, ..., moveFun = coarseLine,
  movePars = list())
}
\arguments{
\item{fun}{function to be minimised (or maximised)}

\item{x}{starting parameters}

\item{moveFun}{function for one dimensional move}

\item{movePars}{parameters to pass to moveFun

\code{coordDescent} is a coordinate descent algorothm that repeately applies
a one-dimensional optimization.  This will converge to the local 
optimum if the function is the sum of a convex and
differentiable function and a separable convex penalty (and the one-dimensional
optimization is consistent).  The 1-D optimization function must have
arguments \code{fun} and \code{x}, and parameters are passed via
\code{movePars}.  The default is \code{coarseLine}.}

\item{maximize}{logical indicating max or min}

\item{help}{further arguments to \code{fun}}
}
\description{
Generic function for performing coordinate descent
}
