% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/subsetMatrix.R
\name{designMatrix}
\alias{designMatrix}
\title{Orthogonal Design Matrix}
\usage{
designMatrix(n)
}
\arguments{
\item{n}{integer containing the number of elements in the set.}
}
\value{
An integer matrix of dimension 2^n by 2^n containing 1 and -1.
}
\description{
Produces a matrix whose rows correspond to an orthogonal binary design matrix.
}
\note{
The output matrix has orthogonal columns and is symmetric, so (up to a constant) is its own inverse.
}
\examples{
designMatrix(3)
}
\author{
Robin Evans
}
\seealso{
\code{\link{combinations}}, \code{\link{subsetMatrix}}.
}
\keyword{arith}

