% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetMatrix.R
\name{fastMobius}
\alias{fastMobius}
\alias{invMobius}
\title{Fast Moebius and inverse Moebius transforms}
\usage{
fastMobius(x, pad = FALSE)

invMobius(x, pad = FALSE)
}
\arguments{
\item{x}{vector to transform}

\item{pad}{logical, should vector not of length 2^k be padded with zeroes?}
}
\description{
Uses the fast method of Kennes and Smets (1990) to obtain Moebius and 
inverse Moebius transforms.
}
\details{

}
\section{Functions}{
\itemize{
\item \code{invMobius}: inverse transform
}}

