% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int2set.R
\name{int2set}
\alias{int2set}
\alias{set2int}
\title{Alternate between sets and integers representing sets of integers via bits}
\usage{
int2set(n, index = 1, simplify = FALSE)

set2int(x, index = 1)
}
\arguments{
\item{n}{integer respresenting a set}

\item{index}{integer to start from}

\item{simplify}{logical: return a single list if \code{n} has length 1?}

\item{x}{list of sets}
}
\value{
For \code{int2set} a list of sets one for each integer 
supplied, for \code{set2int} a vector of the same length as the number 
of sets supplied.
}
\description{
Alternate between sets and integers representing sets of integers via bits
}
\details{
Converts an integer into its binary representation and 
interprets this as a set of integers.  Cannot handle sets with more
than 31 elements.
}
\section{Functions}{
\itemize{
\item \code{set2int()}: Convert sets to integers

}}
