% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.subset.R
\name{is.subset}
\alias{is.subset}
\alias{\%subof\%}
\title{Check subset inclusion}
\usage{
is.subset(x, y)

x \%subof\% y
}
\arguments{
\item{x}{vector.}

\item{y}{vector.}
}
\value{
A logical of length 1.
}
\description{
Determines whether one vector contains all the elements of another.
}
\details{
Determines whether or not every element of \code{x} is also found in
\code{y}. Returns \code{TRUE} if so, and \code{FALSE} if not.
}
\section{Functions}{
\itemize{
\item \code{x \%subof\% y}: operator version

}}
\examples{

is.subset(1:2, 1:3)
is.subset(1:2, 2:3)
1:2 \%subof\% 1:3
1:2 \%subof\% 2:3

}
\seealso{
\code{\link{setmatch}}.
}
\author{
Robin Evans
}
\keyword{arith}
