% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.wholenumber.R
\name{is.wholenumber}
\alias{is.wholenumber}
\title{Determine whether number is integral or not.}
\usage{
is.wholenumber(x, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{numeric vector to be tested.}

\item{tol}{The desired precision.}
}
\value{
A logical vector of the same length as \code{x}, containing the
results of the test.
}
\description{
Checks whether a numeric value is integral, up to machine or other specified
prescision.
}
\examples{

x = c(0.5, 1, 2L, 1e-20)
is.wholenumber(x)

}
\author{
Robin Evans
}
\keyword{arith}

