% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetMatrix.R
\name{kronPower}
\alias{kronPower}
\title{Kronecker power of a matrix or vector}
\usage{
kronPower(x, n)
}
\arguments{
\item{x}{matrix or vector}

\item{n}{integer containing power to take}
}
\description{
Kronecker power of a matrix or vector
}
\details{
This computes \code{x \%x\% ... \%x\% x} 
for \code{n} 
instances of \code{x}.
}
