\name{rje-package}
\alias{rje-package}
\alias{rje}
\docType{package}
\title{
Miscellaneous useful functions
}
\description{
A series of useful functions, some available in different
forms in other packages, but which have been extended, sped up, or
otherwise modified in some way considered useful to the author.
}
\details{
\tabular{ll}{
Package: \tab rje\cr
Type: \tab Package\cr
Version: \tab 1.9\cr
Date: \tab 2014-04-22\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
}
\author{
Robin Evans

Mathias Drton

Maintainer: Robin Evans <evans@stats.ox.ac.uk>
}
\keyword{package}
\keyword{array}
\keyword{arith}
\keyword{optimize}
