% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rje.R
\docType{package}
\name{rje}
\alias{rje}
\alias{rje-package}
\title{Miscellaneous useful functions}
\description{
A series of functions in some way considered useful to the author.  
These include functions for subsetting tables and generating indices for 
arrays, conditioning and intervening in probability distributions, 
generating combinations and more...

Mathias Drton

Maintainer: Robin Evans <evans@stats.ox.ac.uk>
}
\details{
\tabular{ll}{
Package: \tab rje\cr
Type: \tab Package\cr
Version: \tab 1.10.16\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
}
\author{
Robin Evans
}
\keyword{arith}
\keyword{array}
\keyword{optimize}
\keyword{package}
