% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rowMins.R
\name{rowMins}
\alias{rowMaxs}
\alias{rowMins}
\title{Row-wise minima and maxima}
\usage{
rowMins(x)
rowMaxs(x)
}
\arguments{
\item{x}{a numeric (or logical) matrix or data frame}
}
\value{
numeric vector of length \code{nrow(x)} giving the row-wise minima (or maxima) of \code{x}.
}
\description{
Row-wise minima and maxima
}
\details{
The function coerces \code{x} to be a data frame and then
uses \code{pmin} (\code{pmax}) on it.  This is the same as
\code{apply(x, 1, min)} but generally faster if the number of rows
is large.
}

