% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rprobdist}
\alias{rprobdist}
\title{Generate a joint or conditional probability distribution}
\usage{
rprobdist(dim, d, alpha = 1)
}
\arguments{
\item{dim}{the joint dimension of the probability table}

\item{d}{number of dimensions}

\item{alpha}{Dirichlet hyper parameter, defaults to 1 (flat density).}
}
\description{
Wrapper functions to quickly generate discrete joint
(or conditional) distributions using Dirichlets
}
\details{
\code{rprobdist} gives an array of dimension \code{dim}
(recycled as necessary to have length \code{d}, if this is
supplied) whose entries are probabilities drawn from a Dirichlet
distribution whose parameter vector has entries equal to
\code{alpha} (appropriately recycled).
}
\examples{
rprobdist(2, 4)     # 2x2x2x2 table
rprobdist(c(2,3,2)) # 2x3x2 table

rprobdist(2, 4, alpha=1/16)     # using unit information prior
}
\author{
Robin Evans
}
\keyword{distribution}

