% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setmatch.R
\name{sets_nested}
\alias{sets_nested}
\title{Check list of sets is nested}
\usage{
sets_nested(x)
}
\arguments{
\item{x}{list containing collection of sets}
}
\value{
If the sets are nested it returns an ordering, otherwise `NA`.
}
\description{
Check list of sets is nested
}
