% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orders.R
\name{subsetOrder}
\alias{subsetOrder}
\title{Compare sets for inclusion.}
\usage{
subsetOrder(x, y)
}
\arguments{
\item{x}{A vector.}

\item{y}{A vector of the same type as \code{x}.}
}
\value{
A single integer, 0, -1 or 1.
}
\description{
A wrapper for \code{is.subset} which returns set inclusions.
}
\details{
If \code{x} is a subset of \code{y}, returns \code{-1}, for the reverse
returns \code{1}.  If sets are equal or incomparable, it returns \code{0}.
}
\examples{

subsetOrder(2:4, 1:4)
subsetOrder(2:4, 3:5)

}
\seealso{
\code{\link{is.subset}}.
}
\author{
Robin Evans
}
\keyword{arith}
