% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/is.subset.R
\name{is.subset}
\alias{\%subof\%}
\alias{is.subset}
\title{Check subset inclusion}
\usage{
is.subset(x, y)
}
\arguments{
\item{x}{vector.}

\item{y}{vector.}
}
\value{
A logical of length 1.
}
\description{
Determines whether one vector contains all the elements of another.
}
\details{
Determines whether or not every element of \code{x} is also found in
\code{y}. Returns \code{TRUE} if so, and \code{FALSE} if not.
}
\examples{
is.subset(1:2, 1:3)
is.subset(1:2, 2:3)
}
\author{
Robin Evans
}
\seealso{
\code{\link{setmatch}}.
}
\keyword{arith}

