% Generated by roxygen2 (4.0.2): do not edit by hand
\name{last}
\alias{last}
\title{Last element of a vector or list}
\usage{
last(x)
}
\arguments{
\item{x}{a list or vector.}
}
\value{
An object of the same type as \code{x} of length 1 (or empty if
\code{x} is empty).
}
\description{
Returns the last element of a list or vector.
}
\details{
Designed to be faster than using \code{tail()} or \code{rev()}, and cleaner
than writing \code{x[length(x)]}.
}
\examples{
last(1:10)
}
\author{
Robin Evans
}
\seealso{
\code{\link[utils]{tail}}, \code{\link[base]{rev}}.
}
\keyword{manip}

