\name{rje-package}
\alias{rje-package}
\alias{rje}
\docType{package}
\title{
Miscellaneous useful functions
}
\description{
A series of useful functions, some available in different
  forms in other packages, but which have been extended or modified in some way considered useful to the author.
}
\details{
\tabular{ll}{
Package: \tab rje\cr
Type: \tab Package\cr
Version: \tab 1.4\cr
Date: \tab 2012-05-15\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
}
\author{
Robin Evans

Mathias Drton

Maintainer: Robin Evans <rje42@cam.ac.uk>
}
\keyword{package}
\keyword{array}
\keyword{arith}
