\name{rje-package}
\alias{rje-package}
\alias{rje}
\docType{package}
\title{
Miscellaneous useful functions
}
\description{
A series of useful functions, some of which are available in other
  forms in other packages, but many of which have been extended in some way.
}
\details{
\tabular{ll}{
Package: \tab rje\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-07-14\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
}
\author{
Robin Evans

Maintainer: Robin Evans <rje42@uw.edu>
}
\keyword{package}
