% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{rje-package}
\alias{rje}
\alias{rje-package}
\title{Miscellaneous useful functions}
\description{
A series of useful functions, some available in different forms in other
packages, but which have been extended, sped up, or otherwise modified in
some way considered useful to the author.
}
\details{
\tabular{ll}{
Package: \tab rje\cr
Type: \tab Package\cr
Version: \tab 1.10\cr
Date: \tab 2014-08-31\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
}
\author{
Robin Evans

Mathias Drton

Maintainer: Robin Evans <evans@stats.ox.ac.uk>
}
\keyword{arith}
\keyword{array}
\keyword{optimize}
\keyword{package}

